+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/jenkinsci/warnings-ng-plugin/compare/warnings-ng-2.1.1...master)
+## [Unreleased](https://github.com/jenkinsci/warnings-ng-plugin/compare/warnings-ng-2.1.2...master)
+
+## [2.1.2](https://github.com/jenkinsci/warnings-ng-plugin/compare/warnings-ng-2.1.1...warnings-ng-2.1.2) - 2019-1-28
+
+### Fixed
+- Fixed sandbox bypass via CSRF 
+(see [Jenkins Security Advisory 2019-01-28](https://jenkins.io/security/advisory/2019-01-28/))
 
 ## [2.1.1](https://github.com/jenkinsci/warnings-ng-plugin/compare/warnings-ng-2.1.0...warnings-ng-2.1.1) - 2019-1-21